+++ ChangeLog.rst
@@ -24,6 +24,10 @@ Backwards incompatible changes
   ``foo`` uses ``/accounts/oidc/foo/login/`` as its login URL. Set it to empty
   (``""``) to keep the previous URL structure (``/accounts/foo/login/``).
 
+- The SAML default attribute mapping for ``uid`` has been changed to only
+  include ``urn:oasis:names:tc:SAML:attribute:subject-id``. If the SAML response
+  does not contain that, it will fallback to use ``NameID``.
+
 
 0.59.0 (2023-12-13)
 *******************